requirements.txt
torch>=2.1.0
transformers>=4.36.0
numpy>=1.24.0
einops>=0.7.0
wandb>=0.16.0
datasets>=2.16.0
tiktoken>=0.5.2
accelerate>=0.25.0
safetensors>=0.4.1
tqdm>=4.66.0
pyyaml>=6.0
sentencepiece>=0.1.99
